@@ -1,3 +1,3 @@
 module github.com/zalf-rpm/build-pipeline/hpc/simplace/splitsimplaceproj
 
-go 1.23.4
+go 1.25
