@@ -1,3 +1,3 @@
 module local
 
-go 1.22.4
+go 1.25
